Atualizações feitas hoje. Para compactação de arquivos maiores.

+++ Huffman/compactar.h
@@ -21,6 +21,6 @@ void escrever_cabecalho_inicio(unsigned short cabecalho_inicial, FILE* arquivo);
 void escrever_arvore(Node *cabeca, FILE *arquivo);
 
 /* ESCREVE O TEXTO JÁ COMPACTADO NO ARQUIVO */
-void escrever_texto(unsigned short int *array_texto, int tamanho, FILE *arquivo);
+void escrever_texto(unsigned short int *array_texto, unsigned long long int tamanho, FILE *arquivo);
 
-void criar_array_binarios(Hashtable *ht, unsigned char *txt, int tam_txt, unsigned short *array_binario, unsigned int freq_x_profundidade);
+void criar_array_binarios(Hashtable *ht, unsigned char *txt, unsigned long long int tam_txt, unsigned short *array_binario, unsigned long long int freq_x_profundidade);
